doit: Split out cmd_write
